backend/requirements.txt
fastapi==0.103.2
uvicorn==0.29.0
sqlalchemy==2.0.29
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pymongo==4.6.3
motor==3.4.0
python-dotenv
python-jose

